requirements.txt
asgiref==3.8.1
astroid==3.3.9
awsebcli==3.21.0
backports.tarfile==1.2.0
blessed==1.20.0
boto3==1.37.1
botocore==1.37.4
cement==2.10.14
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
dill==0.3.9
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docutils==0.21.2
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
jmespath==1.0.1
keyring==25.6.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.6.0
nh3==0.2.21
packaging==24.2
pathspec==0.10.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pylint==3.3.5
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
readme_renderer==44.0
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
s3transfer==0.11.2
SecretStorage==3.3.3
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.3
symptom_analysis==1.0.1
termcolor==2.5.0
tomli==2.2.1
tomlkit==0.13.2
twine==6.1.0
typing_extensions==4.12.2
urllib3==1.26.20
watchtower==3.4.0
wcwidth==0.2.13
zipp==3.21.0
